pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "response-mcp-server"
version = "0.1.0"
description = "MCP Server for Microsoft Defender XDR response actions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "msal>=1.28.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.38.0",
    "uvicorn>=0.30.0",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]

[project.scripts]
response-mcp = "response_mcp.server_sse:main"

[tool.hatch.build.targets.wheel]
packages = ["response_mcp"]
